spread out the 'by id' methods to have int, long, Guid, and string overloads. Closes GH-424

+++ src/Marten.Testing/persist_and_load_documents_with_long_ids_Tests.cs
@@ -58,7 +58,7 @@ namespace Marten.Testing
 
             using (var session = theStore.OpenSession())
             {
-                session.Delete<LongDoc>(LongDoc.Id);
+                session.Delete<LongDoc>((int) LongDoc.Id);
                 session.SaveChanges();
             }
